Delete unused files and update imports
