+++ src/components/portfolio/MobileAboutMe.tsx
@@ -2,12 +2,14 @@
 
 import CarouselComponent from '@/components/ui/Carousel/Carousel';
 import CarouselCard from '@/components/ui/Carousel/CarouselCard';
-import aboutMeData from '@/data/aboutMeData';
+import { CONSTANTS } from '@/lib/appConstants';
 
 const MobileAboutMeSection = () => {
+  const aboutMeDetails = CONSTANTS.PORTFOLIO.ABOUT_ME;
+
   return (
     <CarouselComponent>
-      {aboutMeData.map((card) => (
+      {aboutMeDetails.map((card) => (
         <CarouselCard
           key={card.title}
           src={card.src}
