--- src/data/aboutMeData.ts
@@ -1,28 +1,0 @@
-const aboutMeData = [
-  {
-    src: '/me.png',
-    alt: 'Portrait of me',
-    title: 'My Story',
-    description: 'Discover my journey in the tech world',
-  },
-  {
-    src: '/jobs/tesla.png',
-    alt: 'Tesla Icon',
-    title: 'Tesla',
-    description: 'Click to explore my internship and full-time experience',
-  },
-  {
-    src: '/jobs/jpmorgan-chase.png',
-    alt: 'JPMorgan Chase & Co. Icon',
-    title: 'JPMorgan Chase & Co.',
-    description: 'Click to see my internship journey at Chase',
-  },
-  {
-    src: '/jobs/seo-tech.png',
-    alt: 'SEO Tech Developer Icon',
-    title: 'SEO Tech Internship',
-    description: 'Where my tech journey began',
-  },
-];
-
-export default aboutMeData;
