--- src/data/blogData.ts
@@ -1,12 +1,0 @@
-const blogData = {
-  contact: {
-    title: 'Contact Me',
-    content: "Let's collaborate on the next big thing",
-  },
-  blogs: {
-    title: 'Blogs',
-    content: 'Explore my tech thoughts on Medium',
-  },
-};
-
-export default blogData;
