--- src/data/responsiveData.ts
@@ -1,8 +1,0 @@
-const responsiveData = {
-  mobile: {
-    breakpoint: { max: 1000, min: 0 },
-    items: 1,
-  },
-};
-
-export default responsiveData;
