+++ src/lib/networkUtils.ts
